requirements.txt
tf-keras
numpy==1.26.4
pandas>=2.0.0
pillow
opencv-python-headless
streamlit>=1.32.0